requirements.txt
build==0.10.0
click==8.1.3
colorama==0.4.5
ez-setup==0.9
Flask==2.2.2
Flask-MySQLdb==1.0.1
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
mariadb==1.1.4
MarkupSafe==2.1.1
mysqlclient==2.1.1
packaging==23.1
pyproject_hooks==1.0.0
python-dotenv==0.21.0
tomli==2.0.1
Werkzeug==2.2.2
zipp==3.9.0
